Get subgraph synced status in SDK
