+++ packages/sdk/src/index.ts
@@ -1,3 +1,3 @@
-export * from "./sdk";
-export * from "./subgraph";
-export { getDeployedTransactionManagerContract } from "./transactionManager";
+export { NxtpSdk } from "./sdk";
+export { getDeployedTransactionManagerContract } from "./transactionManager/transactionManager";
+export * from "./types";
