+++ packages/sdk/src/subgraph/queries.ts
@@ -169,3 +169,13 @@ export const getTransactionsByIdsQuery = gql`
     }
   }
 `;
+
+export const getBlockNumber = gql`
+  query GetBlockNumber {
+    _meta {
+      block {
+        number
+      }
+    }
+  }
+`;
